refactor: Remove dependency on elm-ui
In the end, I'll need to learn CSS anyways and there are things that are
seemingly harder to do with elm-ui. So I'm just committing to finally
learning the ropes and putting in the work to learn the dang styling
language.

Added a color theme as well that switches based on the user's preference
of color mode. It's still based on rose-pine and the canvas still needs
to get ported over to using this as well.